Buck regulator Schematic
Added type - I Compensator. It solved the issue of oscillations in error amplifier (Which induced PSM mode in Regulator). The oscillation was observed when the error amplifier is used in open loop mode with buck regulator in closed loop. The output ripple voltage got amplified, producing ripple (oscillation) at the output of the error amplifier. This PSM Mode was still giving regulated output but deviates from the PWM theory.
Moving to type-3 compensator as type - 1 is limiting the bandwidth.